Extract DataFieldFee composable
