+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/eip20approve/Eip20ApproveConfirmFragment.kt
@@ -1,22 +1,15 @@
 package io.horizontalsystems.bankwallet.modules.eip20approve
 
 import android.os.Parcelable
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalView
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
@@ -24,16 +17,14 @@ import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseComposeFragment
 import io.horizontalsystems.bankwallet.core.setNavigationResultX
-import io.horizontalsystems.bankwallet.core.slideFromBottom
 import io.horizontalsystems.bankwallet.core.slideFromRight
 import io.horizontalsystems.bankwallet.modules.confirm.ConfirmTransactionScreen
 import io.horizontalsystems.bankwallet.modules.eip20approve.AllowanceMode.OnlyRequired
 import io.horizontalsystems.bankwallet.modules.eip20approve.AllowanceMode.Unlimited
 import io.horizontalsystems.bankwallet.modules.evmfee.Cautions
-import io.horizontalsystems.bankwallet.modules.evmfee.FeeSettingsInfoDialog
-import io.horizontalsystems.bankwallet.modules.multiswap.QuoteInfoRow
 import io.horizontalsystems.bankwallet.modules.multiswap.TokenRow
 import io.horizontalsystems.bankwallet.modules.multiswap.TokenRowUnlimited
+import io.horizontalsystems.bankwallet.modules.multiswap.ui.DataFieldFee
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.ButtonPrimaryDefault
 import io.horizontalsystems.bankwallet.ui.compose.components.ButtonPrimaryYellow
@@ -42,8 +33,6 @@ import io.horizontalsystems.bankwallet.ui.compose.components.TransactionInfoCont
 import io.horizontalsystems.bankwallet.ui.compose.components.VSpacer
 import io.horizontalsystems.bankwallet.ui.compose.components.cell.BoxBorderedTop
 import io.horizontalsystems.bankwallet.ui.compose.components.cell.SectionUniversalLawrence
-import io.horizontalsystems.bankwallet.ui.compose.components.subhead2_grey
-import io.horizontalsystems.bankwallet.ui.compose.components.subhead2_leah
 import io.horizontalsystems.core.SnackbarDuration
 import io.horizontalsystems.core.helpers.HudHelper
 import kotlinx.coroutines.delay
@@ -162,42 +151,10 @@ fun Eip20ApproveConfirmScreen(navController: NavController) {
 
         VSpacer(height = 16.dp)
         SectionUniversalLawrence {
-            QuoteInfoRow(
-                title = {
-                    val title = stringResource(id = R.string.FeeSettings_NetworkFee)
-                    val infoText = stringResource(id = R.string.FeeSettings_NetworkFee_Info)
-
-                    subhead2_grey(text = title)
-
-                    Image(
-                        modifier = Modifier
-                            .padding(horizontal = 8.dp)
-                            .clickable(
-                                onClick = {
-                                    navController.slideFromBottom(
-                                        R.id.feeSettingsInfoDialog,
-                                        FeeSettingsInfoDialog.Input(title, infoText)
-                                    )
-                                },
-                                interactionSource = remember { MutableInteractionSource() },
-                                indication = null
-                            )
-                        ,
-                        painter = painterResource(id = R.drawable.ic_info_20),
-                        contentDescription = ""
-                    )
-
-                },
-                value = {
-                    val primary = uiState.networkFee?.primary?.getFormattedPlain() ?: "---"
-                    val secondary = uiState.networkFee?.secondary?.getFormattedPlain() ?: "---"
-
-                    Column(horizontalAlignment = Alignment.End) {
-                        subhead2_leah(text = primary)
-                        VSpacer(height = 1.dp)
-                        subhead2_grey(text = secondary)
-                    }
-                }
+            DataFieldFee(
+                navController,
+                uiState.networkFee?.primary?.getFormattedPlain() ?: "---",
+                uiState.networkFee?.secondary?.getFormattedPlain() ?: "---"
             )
         }
