+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/sendevmtransaction/SendEvmTransactionView.kt
@@ -1,8 +1,5 @@
 package io.horizontalsystems.bankwallet.modules.sendevmtransaction
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -13,7 +10,6 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalView
@@ -27,11 +23,9 @@ import io.horizontalsystems.bankwallet.core.ethereum.CautionViewItem
 import io.horizontalsystems.bankwallet.core.iconPlaceholder
 import io.horizontalsystems.bankwallet.core.imageUrl
 import io.horizontalsystems.bankwallet.core.shorten
-import io.horizontalsystems.bankwallet.core.slideFromBottom
 import io.horizontalsystems.bankwallet.modules.evmfee.Cautions
-import io.horizontalsystems.bankwallet.modules.evmfee.FeeSettingsInfoDialog
-import io.horizontalsystems.bankwallet.modules.multiswap.QuoteInfoRow
 import io.horizontalsystems.bankwallet.modules.multiswap.ui.DataField
+import io.horizontalsystems.bankwallet.modules.multiswap.ui.DataFieldFee
 import io.horizontalsystems.bankwallet.modules.send.SendModule
 import io.horizontalsystems.bankwallet.modules.send.evm.settings.SendEvmNonceViewModel
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
@@ -84,42 +78,10 @@ fun SendEvmTransactionView(
 
         VSpacer(height = 16.dp)
         SectionUniversalLawrence {
-            QuoteInfoRow(
-                title = {
-                    val title = stringResource(id = R.string.FeeSettings_NetworkFee)
-                    val infoText = stringResource(id = R.string.FeeSettings_NetworkFee_Info)
-
-                    subhead2_grey(text = title)
-
-                    Image(
-                        modifier = Modifier
-                            .padding(horizontal = 8.dp)
-                            .clickable(
-                                onClick = {
-                                    navController.slideFromBottom(
-                                        R.id.feeSettingsInfoDialog,
-                                        FeeSettingsInfoDialog.Input(title, infoText)
-                                    )
-                                },
-                                interactionSource = remember { MutableInteractionSource() },
-                                indication = null
-                            )
-                        ,
-                        painter = painterResource(id = R.drawable.ic_info_20),
-                        contentDescription = ""
-                    )
-
-                },
-                value = {
-                    val primary = networkFee?.primary?.getFormattedPlain() ?: "---"
-                    val secondary = networkFee?.secondary?.getFormattedPlain() ?: "---"
-
-                    Column(horizontalAlignment = Alignment.End) {
-                        subhead2_leah(text = primary)
-                        VSpacer(height = 1.dp)
-                        subhead2_grey(text = secondary)
-                    }
-                }
+            DataFieldFee(
+                navController,
+                networkFee?.primary?.getFormattedPlain() ?: "---",
+                networkFee?.secondary?.getFormattedPlain() ?: "---"
             )
         }
